pdf_organizer/requirements.txt
Flask==3.0.3
Werkzeug==3.0.4
pypdf==4.3.1
pdfminer.six==20231228
